assets/python/2025-10-12-agentic-workflows/pyproject.toml
[project]
name = "2025-10-12-agentic-workflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0.11",
    "google-adk>=1.16.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
]
